% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{weight.calculate.ps}
\alias{weight.calculate.ps}
\title{The Inverse Probability Weight (IPW) for Average Treatment Effect (ATE)}
\usage{
weight.calculate.ps(Z, ps, standardize = TRUE)
}
\arguments{
\item{Z}{The binary treatment indicator. A vector with 2 unique numeric values in 0 = untreated and 1 = treated.}

\item{ps}{The propensity score used to calculate the IPW weight. A vector with numeric values greater than 0 and less than 1.
The length of "ps" should equal to the length of "Z".}

\item{standardize}{If standardize the calculated IPW or not. When standardize = FALSE, the weights are calculated
by the IPW definition shown in the description. standardize = TRUE by default, which makes the
the sum of weights be 1 in the untreated/treated group.}
}
\value{
A list containing the following components:
\itemize{
\item{"weight"}: The IPW weights calculated by the input propensity score
\item{"ps"}: The input propensity score
}
}
\description{
This function calculates the IPW for ATE as Tr/ps + (1-Tr)/(1-ps), where Tr is the treatment
indicator and ps is the propensity score.
}
\examples{
KS = Kang_Schafer_Simulation(n = 1000, seeds = 5050)
tr = KS$Data[,2]
true.ps = KS$Data[,11]
true.w = weight.calculate.ps(Z = tr, ps = true.ps, standardize = TRUE)
summary(true.w$weight)
c(sum(true.w$weight[which(tr==0)]), sum(true.w$weight[which(tr==1)]))

}
